.github/workflows/embedding_service/requirements.txt
boto3
pdfminer.six
transformers
torch
pinecone-client
fastapi
uvicorn
pydub
speechrecognition
moviepy
tqdm
scipy
faiss-cpu
spacy
rdflib  # For knowledge graph
pandas